requirements.txt
numpy
scipy
sympy
Django==3.1rc1
django-bootstrap4
django-widget-tweaks
espressodb==1.2.1
networkx
requests
plotly
djangorestframework
